The LEDGroup class can now take any input and determine the number of leds to turn on depending on any set range of input values.
The input, minimum, maximum, and range can now be decimal values.

If the input value is above or below the maximum or minimum value,
respectively, then a catch in the program sets the input equal to
the min or max value. Before doing this, the results from math involving
the input value were unpredictable and unreliable.

The highestLitLedIndex is now always set to -1 if the value is below 0. A
catch in the program does not turn on any leds if the highestLitLedIndex is
set to -1.
In the setAllLedStatesToRepresentInputValue function, all of the leds in
the ledArray are now turned off at the beginning of the function; the
proper number of leds are then turned on. By first turning off all of the
leds, this ensures that no leds are turned on in accordance to the previous
input value and that no leds are turned on if none should be.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>