The program can now adjust the notcherUIs within the mainPanel according to the number of notchers, for up to four notchers.
The index number of each notcherController in the array in the main
controller is passed into both the notcherUI and the notcherController.

mainPanel in view no longer has any ledGroups, textfields, buttons, etc.
added to it directly. All of these are added to the notcherUI, which is
then added to the mainPanel in view.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>