Some minor changes with the changeNameDialog.
The mainFrame is now passed into the constructor of the NotcherUI; this
	allows us to assign a JFrame to the changeNameDialog.
A centerJDialog function has been added to the NotcherUI class.

An apply button has been added to the changeNameDialog.
Some containing panels have been added so that the objects in
	changeNameDialog can be centered.
The size of changeNameDialog has also been adjusted for its content.
The changeNameDialog is also centered after all of the objects have been
	added to the content pane.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>